+++ src/uart.h
@@ -3,9 +3,9 @@
 
 void uart_init ();
 void uart_sendc (char c);
-void uart_sendi (int num);
+void uart_sendi (long num);
 void uart_sendf (float num);
-void uart_sendh (unsigned int num);
+void uart_sendh (unsigned long num);
 void uart_send (char *str);
 char uart_getc ();
 void uart_getline (char *buffer, unsigned int size);
